added HR signature script
